feat: Add firebase analytics
