+++ src/App.js
@@ -1,34 +1,12 @@
 import logo from './logo.svg';
 import './App.css';
+import logAnalytics from './analytics/FirebaseAnalytics.js'
 
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-import { logEvent } from "firebase/analytics";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: "AIzaSyBg4Vvyk-JZvuigcCEikQYOG08k2tbZdwo",
-  authDomain: "wedding-site-106a8.firebaseapp.com",
-  projectId: "wedding-site-106a8",
-  storageBucket: "wedding-site-106a8.appspot.com",
-  messagingSenderId: "641390355622",
-  appId: "1:641390355622:web:f25524a2406d47787f8067",
-  measurementId: "G-2EKR1536ND"
-};
-
-// Initialize Firebase
-const firebaseApp = initializeApp(firebaseConfig);
-const firebaseAnalytics = getAnalytics(firebaseApp);
+console.log("App.js ran")
 
 function App() {
-  function logClick() {
-    console.log("Clicked!")
-    logEvent(firebaseAnalytics, "page_load_main")
-  }
+  logAnalytics("page_render_main")
+
   return (
     <div className="App">
       <header className="App-header">
@@ -44,9 +22,6 @@ function App() {
         >
           Learn React
         </a>
-        <button onClick={logClick}>
-          Click me!
-        </button>
       </header>
     </div>
   );
